extract shared AccessibilityStep component to reduce duplication between Patient and Professional register flows
