+++ src/features/auth/components/ProfissionalRegister/AccessibilityStep/index.tsx
@@ -1,54 +1,7 @@
-"use client";
-import { useState } from "react";
-
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { useProfissionalRegisterStore } from "../useProfissionalRegisterStore";
-import { AccessibilityForm } from "../../AccessibilityShared/AccessibilityForm";
-
-import { ACCESSIBILITY_OPTIONS, accessibilitySchema, AccessibilityData as Data } from "../../AccessibilityShared/constants";
-
+import { AccessibilityStepShared } from "../../AccessibilityShared/AccessibilityStepShared";
 
 export const AccessibilityStep = () => {
-  const { setValue, handleSubmit, watch, getValues } = useForm<Data>({
-    mode: "all",
-    resolver: zodResolver(accessibilitySchema),
-    defaultValues: { accessibility: [] },
-  });
-
-  const { changeStep, updateFields } = useProfissionalRegisterStore();
-  const [expanded, setExpanded] = useState(false);
-  const selected = watch("accessibility");
-
-  const toggleOption = (value: string) => {
-    const cur = getValues("accessibility");
-    const next = cur.includes(value) ? cur.filter((v) => v !== value) : [...cur, value];
-    setValue("accessibility", next, { shouldValidate: true });
-  };
-
-  const onContinue = handleSubmit((data) => {
-    updateFields({ ...(data as any), skippedAccessibility: false });
-    window.scrollTo({ top: 0, behavior: "smooth" });
-    changeStep("complete_profile");
-  });
-
-  const onSkip = () => {
-    updateFields({ skippedAccessibility: true, accessibility: undefined });
-    window.scrollTo({ top: 0, behavior: "smooth" });
-    changeStep("complete_profile");
-  };
-
-  return (
-    <form onSubmit={(e) => e.preventDefault()} className="flex flex-col gap-8">
-      <AccessibilityForm
-        options={ACCESSIBILITY_OPTIONS}
-        selected={selected}
-        onToggleOption={toggleOption}
-        expanded={expanded}
-        onToggleExpand={() => setExpanded((p) => !p)}
-        onSkip={onSkip}
-        onContinue={onContinue}
-      />
-    </form>
-  );
+  const { changeStep: storeChangeStep, updateFields } = useProfissionalRegisterStore();
+  return <AccessibilityStepShared changeStep={(step: string) => storeChangeStep(step as any)} updateFields={updateFields} />;
 };
